+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Open first url in message body
 - Ignore statistic channels in stats
 - Improved layout and keyboard handling for signup
+- Link Getting started in app settings #62
 
 ## 1.3.1